Restructured view folder
