+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:wh_covid19/view/home_page.dart';
-import 'package:wh_covid19/view/ppe_view.dart';
+import 'package:wh_covid19/view/ppe/ppe_view.dart';
 import 'package:wh_covid19/view/sbs_guide_view.dart';
-import 'package:wh_covid19/view/staff_welfare_view.dart';
+import 'package:wh_covid19/view/staff_welfare/staff_welfare_view.dart';
 
 void main() => runApp(MyApp());
